# docs/openapi.json
{"openapi":"3.0.3","info":{"title":"千语博客 API","version":"1.0.0","description":"简单博客系统接口文档，可直接导入 Apifox"},"servers":[{"url":"http://localhost:8080","description":"本地开发环境"}],"tags":[{"name":"公开博客"},{"name":"后台认证"},{"name":"后台分类"},{"name":"后台文章"}],"paths":{"/api/public/categories":{"get":{"tags":["公开博客"],"summary":"获取公开分类列表","responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryListResponse"}}}}}}},"/api/public/posts":{"get":{"tags":["公开博客"],"summary":"获取已发布文章列表","parameters":[{"name":"categoryId","in":"query","schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","schema":{"type":"integer","default":0}},{"name":"size","in":"query","schema":{"type":"integer","default":10}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostPageResponse"}}}}}}},"/api/public/posts/{id}":{"get":{"tags":["公开博客"],"summary":"获取已发布文章详情","parameters":[{"$ref":"#/components/parameters/IdPath"}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponseWrapper"}}}},"404":{"description":"文章不存在"}}}},"/api/admin/auth/login":{"post":{"tags":["后台认证"],"summary":"管理员登录","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"登录成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseWrapper"}}}},"401":{"description":"用户名或密码错误"}}}},"/api/admin/categories":{"get":{"tags":["后台分类"],"summary":"后台分类列表","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryListResponse"}}}},"401":{"description":"未登录"}}},"post":{"tags":["后台分类"],"summary":"新增分类","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRequest"}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseWrapper"}}}},"401":{"description":"未登录"},"409":{"description":"分类冲突"}}}},"/api/admin/categories/{id}":{"put":{"tags":["后台分类"],"summary":"编辑分类","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdPath"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRequest"}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseWrapper"}}}},"401":{"description":"未登录"},"404":{"description":"分类不存在"}}},"delete":{"tags":["后台分类"],"summary":"删除分类","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdPath"}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponse"}}}},"401":{"description":"未登录"},"409":{"description":"分类下存在文章"}}}},"/api/admin/posts":{"get":{"tags":["后台文章"],"summary":"后台文章列表","security":[{"bearerAuth":[]}],"parameters":[{"name":"page","in":"query","schema":{"type":"integer","default":0}},{"name":"size","in":"query","schema":{"type":"integer","default":10}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostPageResponse"}}}},"401":{"description":"未登录"}}},"post":{"tags":["后台文章"],"summary":"新增文章","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRequest"}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponseWrapper"}}}},"401":{"description":"未登录"}}}},"/api/admin/posts/{id}":{"get":{"tags":["后台文章"],"summary":"后台文章详情","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdPath"}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponseWrapper"}}}},"401":{"description":"未登录"},"404":{"description":"文章不存在"}}},"put":{"tags":["后台文章"],"summary":"编辑文章","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdPath"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostRequest"}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponseWrapper"}}}},"401":{"description":"未登录"},"404":{"description":"文章不存在"}}},"delete":{"tags":["后台文章"],"summary":"删除文章","security":[{"bearerAuth":[]}],"parameters":[{"$ref":"#/components/parameters/IdPath"}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmptyResponse"}}}},"401":{"description":"未登录"},"404":{"description":"文章不存在"}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"parameters":{"IdPath":{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}},"schemas":{"ApiResponseBase":{"type":"object","properties":{"code":{"type":"integer","example":0},"message":{"type":"string","example":"success"}}},"EmptyResponse":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"nullable":true}}}]},"LoginRequest":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","example":"admin"},"password":{"type":"string","example":"admin123"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"tokenType":{"type":"string","example":"Bearer"},"expiresInMinutes":{"type":"integer","example":120}}},"LoginResponseWrapper":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"$ref":"#/components/schemas/LoginResponse"}}}]},"CategoryRequest":{"type":"object","required":["name","slug"],"properties":{"name":{"type":"string","example":"技术"},"slug":{"type":"string","example":"tech"},"description":{"type":"string","example":"技术文章"}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"slug":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CategoryResponseWrapper":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Category"}}}]},"CategoryListResponse":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Category"}}}}]},"PostStatus":{"type":"string","enum":["DRAFT","PUBLISHED"]},"PostRequest":{"type":"object","required":["title","summary","content","categoryId","status"],"properties":{"title":{"type":"string","example":"第一篇博客"},"summary":{"type":"string","example":"文章摘要"},"content":{"type":"string","example":"文章正文"},"categoryId":{"type":"integer","format":"int64","example":1},"status":{"$ref":"#/components/schemas/PostStatus"}}},"Post":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"summary":{"type":"string"},"content":{"type":"string"},"status":{"$ref":"#/components/schemas/PostStatus"},"category":{"$ref":"#/components/schemas/Category"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"PostResponseWrapper":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"$ref":"#/components/schemas/Post"}}}]},"PostPage":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Post"}},"page":{"type":"integer"},"size":{"type":"integer"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer"}}},"PostPageResponse":{"allOf":[{"$ref":"#/components/schemas/ApiResponseBase"},{"type":"object","properties":{"data":{"$ref":"#/components/schemas/PostPage"}}}]}}}}